Replace long and unsigned long with portable types

+++ src/converter.h
@@ -8,26 +8,23 @@ class Converter
 {
 public:
     static CodeArray convert_codes(
-        unsigned long point_count, unsigned long cut_count,
-        const unsigned long* cut_start, unsigned long subtract = 0);
+        size_t point_count, size_t cut_count, const size_t* cut_start,
+        size_t subtract = 0);
 
     static CodeArray convert_codes_check_closed(
-        unsigned long point_count, unsigned long cut_count,
-        const unsigned long* cut_start, const double* points);
+        size_t point_count, size_t cut_count, const size_t* cut_start,
+        const double* points);
 
     static CodeArray convert_codes_check_closed_single(
-        unsigned long point_count, const double* points);
+        size_t point_count, const double* points);
 
     static OffsetArray convert_offsets(
-        unsigned long offset_count, const unsigned long* start,
-        unsigned long subtract = 0);
+        size_t offset_count, const size_t* start, size_t subtract = 0);
 
     static OffsetArray convert_offsets_nested(
-        unsigned long offset_count, const unsigned long* start,
-        const unsigned long* nested_start);
+        size_t offset_count, const size_t* start, const size_t* nested_start);
 
-    static PointArray convert_points(
-        unsigned long point_count, const double* start);
+    static PointArray convert_points(size_t point_count, const double* start);
 };
 
 #endif // CONTOURPY_CONVERTER_H
